Interfaces can be viewed...
At this point, I believe most of DocumenterPhp itself can be rendered
using the front end for DocumenterLaravel. This codebase has all major
project objects (class, trait, interface), symbols for those objects
(properties and methods), and the means to instantiate an overarching
project (not inherent in phpDocumentor that I’m aware of). Therefore, I
would say it is complete enough for a 0.0.1 release to also broadcast
it out to the broader community of developers.

As far as abstractions go, I think DocumentPhp does pretty well, and
has a solid architectural beginning, in my biased opinion. :)